Move experiment strategies to dedicated experiments module
In succession of the moved `optuna` experiment strategy, the other
existing strategies have been moved to their own file in the
`experiment` module.
