+++ lir/experiments/__init__.py
@@ -1,0 +1,10 @@
+from lir.experiments.base_experiment import Experiment
+from lir.experiments.optuna import OptunaExperiment
+from lir.experiments.predefined import PredefinedExperiment
+
+
+__all__ = [
+    'Experiment',
+    'OptunaExperiment',
+    'PredefinedExperiment',
+]
